val_group: only depend on iProp when needed
Reorganise the dependency on iProp so that we can apply soundness of the
logical relation to get the contextual equivalence at the end. In particular,
the expressions involved in the security games now only depend on the semantic
group and the `clutch_group_struct` providing the implementations, not on the
laws, and thus not on `clutchRGS`.